basic follow camera functionallity
still some details need to be pollished

A. The camera sits at an initial position and doesn't move during Slingshot's aimingMode.
B. After the projectile is launched, the camera follows it (with a little easing to make it feel smoother).
C. As the camera moves up into the air, increase the Camera.orthographicSize to keep the ground in view.
D. When the projectile comes to rest, the camera stops following it and returns to the initial position.